L2-330: Move test
